feat(dashboard): fold auth tab into settings tab
Replaces the standalone authentication tab on MCP server details with an
authentication section inside the settings tab, redirecting legacy
authentication-tab paths to settings#authentication. Also splits the
settings tab monolith into per-section files under tabs/settings/: a
composition root, the SettingsSection shell primitives, and one
self-contained section component per card.
