+++ client/dashboard/src/pages/mcp/x/MCPServerDetailsRouting.test.ts
@@ -1,5 +1,9 @@
 import { describe, expect, it } from "vitest";
-import { activeTabFromPath } from "./MCPServerDetailsRouting";
+import {
+  activeTabFromPath,
+  initialTabFromHash,
+  isLegacyAuthenticationTabPath,
+} from "./MCPServerDetailsRouting";
 
 describe("activeTabFromPath", () => {
   it("returns no tab for the server details route without a tab segment", () => {
@@ -8,7 +12,7 @@ describe("activeTabFromPath", () => {
     ).toBeUndefined();
   });
 
-  it.each(["overview", "authentication", "team-access", "settings"] as const)(
+  it.each(["overview", "team-access", "settings"] as const)(
     "reads the %s tab when the server slug has the same value",
     (tab) => {
       expect(
@@ -41,13 +45,31 @@ describe("activeTabFromPath", () => {
     ).toBe("settings");
   });
 
+  it("does not treat the legacy authentication path as an active tab", () => {
+    expect(
+      activeTabFromPath(
+        "/acme/projects/default/mcp/x/my-server/authentication",
+        "my-server",
+      ),
+    ).toBeUndefined();
+  });
+
+  it("detects the legacy authentication path for redirects", () => {
+    expect(
+      isLegacyAuthenticationTabPath(
+        "/acme/projects/default/mcp/x/my-server/authentication",
+        "my-server",
+      ),
+    ).toBe(true);
+  });
+
   it("matches decoded server slug segments", () => {
     expect(
       activeTabFromPath(
-        "/acme/projects/default/mcp/x/my%20server/authentication",
+        "/acme/projects/default/mcp/x/my%20server/settings",
         "my server",
       ),
-    ).toBe("authentication");
+    ).toBe("settings");
   });
 
   it("returns no tab for an invalid tab segment", () => {
@@ -57,5 +79,16 @@ describe("activeTabFromPath", () => {
         "my-server",
       ),
     ).toBeUndefined();
+  });
+});
+
+describe("initialTabFromHash", () => {
+  it("maps the legacy authentication hash to settings", () => {
+    expect(initialTabFromHash("#authentication", true)).toBe("settings");
+  });
+
+  it("keeps team access behind the RBAC feature flag", () => {
+    expect(initialTabFromHash("#team-access", false)).toBe("overview");
+    expect(initialTabFromHash("#team-access", true)).toBe("team-access");
   });
 });
