+++ client/dashboard/src/pages/mcp/x/tabs/settings/sections/authentication/DeleteRemoteIdentityProviderDialog.tsx
